+++ src/main/java/it/pagopa/ecommerce/payment/methods/infrastructure/rule/AmountLangFilterRule.java
@@ -5,6 +5,8 @@ import it.pagopa.ecommerce.payment.methods.infrastructure.PspRepository;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
 
+import java.math.BigInteger;
+
 @Component
 public class AmountLangFilterRule implements IFilterRule {
     @Override
@@ -14,7 +16,7 @@ public class AmountLangFilterRule implements IFilterRule {
 
     @Override
     public Flux<PspDocument> execute(PspRepository pspRepository, Integer amount, String language, String paymentTypeCode) {
-        return pspRepository.findPspMatchAmountLang(amount, language);
+        return pspRepository.findPspMatchAmountLang(BigInteger.valueOf(amount), language);
     }
 
 }
